+++ src/main/java/dev/muon/medievalorigins/action/ModBientityActionTypes.java
@@ -1,10 +1,7 @@
 package dev.muon.medievalorigins.action;
 
 import dev.muon.medievalorigins.MedievalOrigins;
-import dev.muon.medievalorigins.action.bientity.CastSpellBientityActionType;
-import dev.muon.medievalorigins.action.bientity.RaycastBetweenActionType;
-import dev.muon.medievalorigins.action.bientity.SpellDamageActionType;
-import dev.muon.medievalorigins.action.bientity.TransferItemActionType;
+import dev.muon.medievalorigins.action.bientity.*;
 import io.github.apace100.apoli.action.ActionConfiguration;
 import io.github.apace100.apoli.action.type.BiEntityActionType;
 import io.github.apace100.apoli.registry.ApoliRegistries;
@@ -25,6 +22,7 @@ public class ModBientityActionTypes {
     public static final ActionConfiguration<SpellDamageActionType> SPELL_DAMAGE = register(ActionConfiguration.of(MedievalOrigins.loc("spell_damage"), SpellDamageActionType.DATA_FACTORY));
     public static final ActionConfiguration<TransferItemActionType> TRANSFER_ITEM = register(ActionConfiguration.of(MedievalOrigins.loc("transfer_item"), TransferItemActionType.DATA_FACTORY));
     public static final ActionConfiguration<RaycastBetweenActionType> RAYCAST_BETWEEN = register(ActionConfiguration.of(MedievalOrigins.loc("raycast_between"), RaycastBetweenActionType.DATA_FACTORY));
+    public static final ActionConfiguration<SpellHealActionType> SPELL_HEAL = register(ActionConfiguration.of(MedievalOrigins.loc("spell_heal"), SpellHealActionType.DATA_FACTORY));
 
     public static void register() {}